netapp: refactoring to use the new structure
